Align project with containerd sub-project requirements
- Rename to go-dmverity
- Add containerd governance and contribution guidelines
- Add CI workflows, linting config, and documentation
- Add license headers to all source files
