+++ cmd/go-dmverity/close_linux_test.go
@@ -1,3 +1,19 @@
+/*
+   Copyright The containerd Authors.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
 package main
 
 import (
@@ -5,7 +21,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/ChengyuZhu6/veritysetup-go/pkg/utils"
+	"github.com/containerd/go-dmverity/pkg/utils"
 )
 
 func TestClose(t *testing.T) {
@@ -129,7 +145,7 @@ func TestClose_NonExistentDevice(t *testing.T) {
 	utils.RequireTool(t, "dmsetup")
 
 	name := "vgo-close-nonexistent"
-	cmd := exec.Command("veritysetup-go", "close", name)
+	cmd := exec.Command("go-dmverity", "close", name)
 	out, err := cmd.CombinedOutput()
 	if err == nil {
 		t.Errorf("expected error for non-existent device, got output: %s", string(out))
