+++ cmd/go-dmverity/open.go
@@ -1,3 +1,19 @@
+/*
+   Copyright The containerd Authors.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
 package main
 
 import (
@@ -7,9 +23,9 @@ import (
 	"os"
 	"strings"
 
-	"github.com/ChengyuZhu6/veritysetup-go/pkg/dm"
-	"github.com/ChengyuZhu6/veritysetup-go/pkg/utils"
-	verity "github.com/ChengyuZhu6/veritysetup-go/pkg/verity"
+	"github.com/containerd/go-dmverity/pkg/dm"
+	"github.com/containerd/go-dmverity/pkg/utils"
+	verity "github.com/containerd/go-dmverity/pkg/verity"
 )
 
 func parseOpenArgs(args []string) (*verity.VerityParams, string, string, string, []byte, []string, string, error) {
