+++ go.mod
@@ -1,7 +1,7 @@
-module github.com/ChengyuZhu6/veritysetup-go
+module github.com/containerd/go-dmverity
 
 go 1.24.3
 
 require golang.org/x/sys v0.27.0
 
-require github.com/google/uuid v1.6.0 // indirect
+require github.com/google/uuid v1.6.0
