+++ pkg/utils/test_util.go
@@ -1,3 +1,19 @@
+/*
+   Copyright The containerd Authors.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
 package utils
 
 import (
@@ -41,7 +57,7 @@ func RunCmd(t *testing.T, name string, args ...string) (string, string) {
 
 func RunGoCLI(t *testing.T, args ...string) (string, string) {
 	t.Helper()
-	bin := RequireTool(t, "veritysetup-go")
+	bin := RequireTool(t, "go-dmverity")
 	ctx, cancel := context.WithTimeout(context.Background(), 120*time.Second)
 	defer cancel()
 	cmd := exec.CommandContext(ctx, bin, args...)
