+++ tests/verity-compat.sh
@@ -1,8 +1,23 @@
 #!/bin/bash
+
+#   Copyright The containerd Authors.
+
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+
+#       http://www.apache.org/licenses/LICENSE-2.0
+
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+
 # Copied from https://gitlab.com/cryptsetup/cryptsetup/-/blob/dc2251b88d55b349793bd89e7ba765b196cd5c34/tests/verity-compat-test
 # with some adjustments.
 
-VERITYSETUP=${VERITYSETUP:-/usr/local/bin/veritysetup-go}
+VERITYSETUP=${VERITYSETUP:-/usr/local/bin/go-dmverity}
 
 DEV_NAME=verity3273
 DEV_NAME2=verity3273x
